# Conflicts:
#	src/ctool1cd/cunicode.cpp

+++ src/SystemClasses/TFileStream.cpp
@@ -60,13 +60,13 @@ void TFileStream::Close()
 	_stream.reset();
 }
 
-size_t TFileStream::Read(void *Buffer, size_t Count)
+int64_t TFileStream::Read(void *Buffer, int64_t Count)
 {
 	Reopen();
 	return TWrapperStream::Read(Buffer, Count);
 }
 
-size_t TFileStream::Write(const void *Buffer, size_t Count)
+int64_t TFileStream::Write(const void *Buffer, int64_t Count)
 {
 	Reopen();
 	return TWrapperStream::Write(Buffer, Count);
